Fix GET /?locale=nl calling a dummy action
The setup process for this plug requires adding a dummy action for GET
"/", and promises that it will never be called. However, it was called
in one specific case, for requests to GET "/?locale=l" where "l" was a
supported locale. This commit ensures that requests to GET "/?locale=l"
where "l" is a valid locale will end up in a redirect to "/l".